// src/app/interfaces/company-profile.ts
export interface CompanyProfile {
  id: number;
  legalName: string;
  tradeName?: string;
  document?: string;
  pixKey?: string;
  email?: string;
  gmailPassword?: string;
  phone?: string;
  whatsapp?: string;
  address?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  instagramLogin?: string;
  instagramPassword?: string;
  contractTerms?: string;
  createdAt?: string;
  updatedAt?: string;
}